Add mobile screenshot script
