+++ mobile/src/components/organisms/AnimatedList.tsx
@@ -8,6 +8,17 @@ import { listItemEnter, listItemExit, listLayout } from "@/src/design";
 // FlatListProps<T> and treat the inner animated list as untyped.
 const RNAnimatedFlatList = Animated.FlatList as unknown as ComponentType<Record<string, unknown>>;
 
+type AnimatedListProps<T> = FlatListProps<T> & {
+  /**
+   * Animate neighbors gliding into place on layout changes (Reanimated
+   * `itemLayoutAnimation`). Default true. Set false for lists whose rows change
+   * height *in place* (e.g. the checklist's expand/collapse): the layout
+   * animation doesn't reflow sibling rows when an item grows, leaving them
+   * overlapping on the New Architecture. Insert/remove animations are unaffected.
+   */
+  animateItemLayout?: boolean;
+};
+
 /**
  * FlatList with built-in motion: each item fades/slides in on insert, out on
  * removal, and neighbors glide into place on layout changes. Drop-in for
@@ -17,7 +28,11 @@ const RNAnimatedFlatList = Animated.FlatList as unknown as ComponentType<Record<
  * renders immediately (no N simultaneous animations on a long list) and only
  * rows inserted afterwards animate in.
  */
-export function AnimatedList<T>({ renderItem, ...props }: FlatListProps<T>) {
+export function AnimatedList<T>({
+  renderItem,
+  animateItemLayout = true,
+  ...props
+}: AnimatedListProps<T>) {
   const [animateInsertions, setAnimateInsertions] = useState(false);
 
   useEffect(() => {
@@ -28,7 +43,7 @@ export function AnimatedList<T>({ renderItem, ...props }: FlatListProps<T>) {
   return (
     <RNAnimatedFlatList
       {...props}
-      itemLayoutAnimation={listLayout()}
+      itemLayoutAnimation={animateItemLayout ? listLayout() : undefined}
       renderItem={(info: ListRenderItemInfo<T>) => (
         <Animated.View
           entering={animateInsertions ? listItemEnter() : undefined}
